Refactor graphql schema and resolver
